Update to v098r05 release.
byuu says:

Changelog:
- WS/WSC: re-added support for screen rotation (code is inside WS core)
- ruby: changed sample(uint16_t left, uint16_t right) to sample(int16_t
  left, int16_t right);
  - requires casting to uint prior to shifting in each driver, but
    I felt it was misleading to use uint16_t just to avoid that
- ruby: WASAPI is now built in by default; has wareya's improvements,
  and now supports latency adjust
- tomoko: audio settings panel has new "Exclusive Mode" checkbox for
  WASAPI driver only
  - note: although the setting *does* take effect in real-time, I'd
    suggest restarting the emulator after changing it
- tomoko: audio latency can now be set to 0ms (which in reality means
  "the minimum supported by the driver")
- all: increased cothread size from 512KiB to 2MiB to see if it fixes
  bullshit AMD driver crashes
  - this appears to cause a slight speed penalty due to cache locality
    going down between threads, though

+++ ruby/audio/openal.cpp
@@ -66,8 +66,8 @@ struct AudioOpenAL : Audio {
     return false;
   }
 
-  auto sample(uint16_t left, uint16_t right) -> void {
-    buffer.data[buffer.length++] = left << 0 | right << 16;
+  auto sample(int16_t left, int16_t right) -> void {
+    buffer.data[buffer.length++] = (uint16_t)left << 0 | (uint16_t)right << 16;
     if(buffer.length < buffer.size) return;
 
     ALuint albuffer = 0;
